+++ examples/basic_containers.rs
@@ -4,8 +4,8 @@ use stacked_errors::{ensure, ensure_eq, Result, StackableErr};
 use super_orchestrator::{
     docker::{Container, ContainerNetwork, Dockerfile},
     net_message::wait_for_ok_lookup_host,
-    FileOptions,
 };
+use tracing::info;
 
 const BASE_CONTAINER: &str = "fedora:40";
 const TIMEOUT: Duration = Duration::from_secs(300);
@@ -19,83 +19,94 @@ async fn main() -> Result<()> {
         .init();
     let logs_dir = "./logs";
 
-    println!("\n\nexample 0\n");
+    info!("\n\nexample 0\n");
 
-    // a default container configuration with the fedora:38 image
-    let container = Container::new("container0", Dockerfile::name_tag(BASE_CONTAINER));
+    // a default container configuration with the `BASE_CONTAINER` image
+    let container = Container::new("example0", Dockerfile::name_tag(BASE_CONTAINER));
 
-    // solo run `/usr/bin/ls -a /` inside the container
-    let comres = container
-        .entrypoint("/usr/bin/ls", ["-a", "/"])
-        .run(None, TIMEOUT, logs_dir, true)
-        .await
-        .stack()?;
+    // set log files to be generated in `logs_dir` and for the entrypoint and
+    // entrypoint args of the container to be equivalent to running "ls -a /" from a
+    // shell inside the container
+    let container = container.log(true).entrypoint("/usr/bin/ls", ["-a", "/"]);
+
+    // run the container in debug mode
+    let comres = container.run(None, TIMEOUT, logs_dir, true).await.stack()?;
     // and get the output as if it were run locally
     comres.assert_success().stack()?;
-    dbg!(&comres.stdout_as_utf8().stack()?);
+    dbg!(comres.stdout_as_utf8().stack()?);
+    ensure!(!comres.stdout.is_empty());
+    ensure!(comres.stderr.is_empty());
 
-    println!("\n\nexample 1\n");
+    info!("\n\nexample 1\n");
+
+    // test that stderr and errors are handled correctly
+    let comres = Container::new("example1", Dockerfile::name_tag(BASE_CONTAINER))
+        .log(true)
+        .entrypoint("/usr/bin/ls", ["-a", "/nonexistent"])
+        .run(None, TIMEOUT, logs_dir, false)
+        .await
+        .stack()?;
+    ensure!(!comres.successful());
+    dbg!(comres.stderr_as_utf8().stack()?);
+    ensure!(comres.stdout.is_empty());
+    ensure!(!comres.stderr.is_empty());
+
+    info!("\n\nexample 2\n");
 
     // sleep for 1 second
-    Container::new("container0", Dockerfile::name_tag(BASE_CONTAINER))
+    Container::new("example2", Dockerfile::name_tag(BASE_CONTAINER))
         .entrypoint("/usr/bin/sleep", ["1"])
-        .run(None, TIMEOUT, logs_dir, true)
+        .run(None, TIMEOUT, logs_dir, false)
         .await
         .stack()?
         .assert_success()
         .stack()?;
 
-    println!("\n\nexample 2\n");
+    info!("\n\nexample 3\n");
 
     // purposely timeout
-    let comres = Container::new("container0", Dockerfile::name_tag(BASE_CONTAINER))
+    let comres = Container::new("example3", Dockerfile::name_tag(BASE_CONTAINER))
         .entrypoint("/usr/bin/sleep", ["infinity"])
-        .run(None, Duration::from_secs(1), logs_dir, true)
+        .run(None, Duration::from_secs(1), logs_dir, false)
         .await;
     dbg!(&comres);
     ensure!(comres.unwrap_err().is_timeout());
 
-    println!("\n\nexample 3\n");
+    info!("\n\nexample 4\n");
 
     // read from a local folder that is mapped to the container's filesystem with a
     // volume
-    let comres = Container::new("container0", Dockerfile::name_tag(BASE_CONTAINER))
+    let comres = Container::new("example4", Dockerfile::name_tag(BASE_CONTAINER))
         .entrypoint("/usr/bin/cat", ["/dockerfile_resources/example.txt"])
         .volume(
             "./dockerfiles/dockerfile_resources/",
             "/dockerfile_resources/",
         )
-        .run(None, TIMEOUT, logs_dir, true)
+        .run(None, TIMEOUT, logs_dir, false)
         .await
         .stack()?;
     comres.assert_success().stack()?;
-    ensure_eq!(
-        comres.stdout_as_utf8().stack()?,
-        FileOptions::read_to_string("./dockerfiles/dockerfile_resources/example.txt")
-            .await
-            .stack()?
-    );
+    ensure_eq!(comres.stdout_as_utf8().stack()?, "hello from example.txt");
 
-    println!("\n\nexample 4\n");
+    info!("\n\nexample 5\n");
 
     // for more complicated things we need `ContainerNetwork`s
     let mut cn = ContainerNetwork::new("test", None, logs_dir);
     cn.add_container(
-        Container::new("container0", Dockerfile::name_tag(BASE_CONTAINER))
+        Container::new("example5", Dockerfile::name_tag(BASE_CONTAINER))
             .entrypoint("/usr/bin/sleep", ["3"]),
     )
     .stack()?;
     // run all containers
     cn.run_all().await.stack()?;
 
-    let uuid = cn.uuid_as_string();
-    // when communicating inside a container to another container in the network,
-    // you would use a hostname like this
-    let host = format!("container0_{uuid}");
-    dbg!(&host);
-    // but outside we need the IP address
+    // when communicating inside a container to another container in the same
+    // network, you can use the `container_name` of the container as the
+    // hostname, in this case "example5"
+
+    // but outside the network we need the IP address
     let host_ip = cn
-        .wait_get_ip_addr(20, Duration::from_millis(300), "container0")
+        .wait_get_ip_addr(20, Duration::from_millis(300), "example5")
         .await
         .stack()?;
     dbg!(&host_ip);